Argument Mining
All files are related to argument mining. Claims and evidence are the main components of an argument; identifying and using them correctly are essential to framing an argument in a debate. Developing machine learning techniques to mine massive corpora for claims and evidence and use them to generate arguments relevant to a controversial topic.