Stance Classification and Sentiment Analysis
An automatic debating system must be able to identify whether an argument supports or contests a given topic. This is fairly easy for humans but difficult for machines, as it requires great sensitivity to the rich subtleties and nuances of natural language. We have made important progress in this intriguing line of research.